CLASS AND ID SELECTORS
An exercise on the correct usage of classes and id's.
Added same background and font family for three elements
sharing the same class and one of them using a second one to Add
another css propertie in addition the the ones shared by the
other two, other element were id's with unique css properties.